Move `RefValue` to be with the rest of the deserialization logic. Add example implementations for `DocumentDeserialize` to existing types
