+++ src/schema/mod.rs
@@ -130,7 +130,7 @@ use columnar::ColumnType;
 
 pub use self::bytes_options::BytesOptions;
 pub use self::date_time_options::{DateOptions, DateTimePrecision, DATE_TIME_PRECISION_INDEXED};
-pub use self::document::{DocParsingError, Document, OwnedValue, TantivyDocument, Value};
+pub use self::document::{DocParsingError, Document, OwnedValue, RefValue, TantivyDocument, Value};
 pub(crate) use self::facet::FACET_SEP_BYTE;
 pub use self::facet::{Facet, FacetParseError};
 pub use self::facet_options::FacetOptions;
